requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.14
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
psycopg2==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
